Rename drops.ts as drop.ts
